refactor(rds_custom toolset): Group toolsets by read/write intent
